Minor fixes to allow compilation under Visual C++.
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/trunk@1025 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/material_manager.hpp
@@ -42,7 +42,7 @@ public:
     Material *getMaterial    (ssgLeaf *lf);
     Material *getMaterial    (const char *texname);
     Material *getMaterial    (const std::string& t) {return getMaterial(t.c_str());}
-    int       getNumEntities ()                     {return m_materials.size();      }
+    int       getNumEntities ()                     {return (int)m_materials.size();      }
 };
 
 extern ssgState *fuzzy_gst, *herringbones_gst;
